upload-service/requirements.txt
fastapi
python-dotenv
jwt
requests